Moved all models in musit-service to new module called musit-models. Removed unecessary Dto for Group.
