+++ musit-models/src/main/scala/no/uio/musit/models/MuseumIdentifier.scala
