+++ musit-models/src/main/scala/no/uio/musit/models/MuseumNo.scala
