+++ musit-models/src/main/scala/no/uio/musit/models/MusitUUID.scala
