+++ musit-models/src/main/scala/no/uio/musit/models/NamedPathElement.scala
