+++ musit-models/src/test/scala/no/uio/musit/models/NodePathSpec.scala
