+++ service_auth/app/models/Group.scala
@@ -19,6 +19,7 @@
 
 package models
 
+import no.uio.musit.models.GroupId
 import no.uio.musit.security.Permissions.Permission
 import play.api.libs.json.{Format, Json, Reads}
 
@@ -33,6 +34,8 @@ object Group {
 
   implicit def format: Format[Group] = Json.format[Group]
 
+  def fromGroupAdd(gid: GroupId, ga: GroupAdd): Group =
+    Group(gid, ga.name, ga.permission, ga.description)
 }
 
 case class GroupAdd(
